main/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
apscheduler
requests
pydantic
tensorflow
